Added shared library and added dependecies on shared
Add Shared project and references to modules

Updated `Basket`, `Catalog`, and `Ordering` project files to include a reference to the new `Shared` project. Created `Shared.csproj` with necessary properties and added a basic `Class1` class. Updated the solution file to include the `Shared` project and its configurations for Debug and Release modes.